examples/robot_localization_failure/pyproject.toml
[project]
name = "robot-localization-failure"
version = "0.1.0"
requires-python = ">=3.10, <4.0"
dependencies = [
  "flowcean",
  "matplotlib>=3.10.0",
  "opencv-python>=4.11.0.86",
  "pillow>=11.1.0",
  "numba>=0.61.0",
  "pyqt6>=6.9.0",
  "plotly>=6.0.1",
  "dash>=3.0.4",
  "optuna>=4.5.0",
  "polars-u64-idx>=1.32.3",
  "catboost>=1.2.8",
  "lightgbm>=4.6.0",
]

[tool.pyright]
extends = "../../pyproject.toml"

[tool.uv.sources]
flowcean = { workspace = true }
